// frontend/src/app/shared/generated/models/analysis-template-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AnalysisTemplateDto {
  calculationDefinitions?: (string | null);
  id: (number | string);
  isRetired: boolean;
  maxTolerance?: (number | string | null);
  minTolerance?: (number | string | null);
  name: string;
  rowVersion: string;
  site: string;
  testConfiguration?: (string | null);
  validationRules?: (string | null);
  version: (number | string);
}